Don't skip orgs
It looks like you're explicitly removing the org and plugin_setting stables and preventing them from being migrated. Why? The org seems to contain important data in most cases. Removing this rm  should fix https://github.com/haron/grafana-migrator/issues/2. 

I'm imagining some default org stuff might be setup as part of the migrations when grafana starts. If that's the case, it seems that the best solution would be to change the insert for the org item with id 1 to be an update. I can work on that if you agree.